As a logged-in user, I want to find a wizard by id, so that I can learn the details of this specific wizard.
