--- src/main/java/org/alpha/omega/hogwarts_artifacts_online/artifact/constant/Constant.java
@@ -1,19 +1,0 @@
-package org.alpha.omega.hogwarts_artifacts_online.artifact.constant;
-
-public class Constant {
-
-    private Constant() {}
-
-    public static class CustomExMessage {
-
-        private CustomExMessage() {}
-
-        public static class Artifact {
-
-            private Artifact() {}
-
-            public static final String INVALID_ARGUMENTS = "Provided arguments are invalid, see data for details.";
-            public static final String NOT_FOUNT_ARTIFACT = "Could not find artifact with Id: %s";
-        }
-    }
-}
