--- src/main/java/org/alpha/omega/hogwarts_artifacts_online/artifact/entity/Artifact.java
@@ -1,64 +1,0 @@
-package org.alpha.omega.hogwarts_artifacts_online.artifact.entity;
-
-import jakarta.persistence.*;
-import lombok.*;
-
-import java.io.Serializable;
-import java.util.Objects;
-
-@Entity(name = "Artifact")
-@Table(name = "Artifacts")
-@Getter
-@Setter
-@NoArgsConstructor
-@AllArgsConstructor
-@Builder
-public class Artifact implements Serializable {
-
-    @Id
-    private String id;
-
-    @Column(name = "name", nullable = false, length = 80)
-    private String name;
-
-    @Column(name = "description", nullable = false, length = 150)
-    private String description;
-
-    @Column(name = "image_url", nullable = false)
-    private String imageUrl;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "wizard_id", referencedColumnName = "id", foreignKey = @ForeignKey(name = "wizard_id_fk"))
-    private Wizard wizard;
-
-    @Override
-    public boolean equals(Object obj) {
-        if (this == obj)
-            return true;
-        if (obj == null)
-            return false;
-        if (getClass() != obj.getClass())
-            return false;
-        Artifact other = (Artifact) obj;
-        if (id == null) {
-            return false;
-        } else if (!id.equals(other.id))
-            return false;
-        return true;
-    }
-
-    @Override
-    public int hashCode() {
-        return Objects.hashCode(id);
-    }
-
-    @Override
-    public String toString() {
-        return "Artifact{" +
-                "id=" + id +
-                ", name='" + name + '\'' +
-                ", description='" + description + '\'' +
-                ", imageUrl='" + imageUrl + '\'' +
-                '}';
-    }
-}
