--- src/main/java/org/alpha/omega/hogwarts_artifacts_online/artifact/exception/NotFoundException.java
@@ -1,8 +1,0 @@
-package org.alpha.omega.hogwarts_artifacts_online.artifact.exception;
-
-public class NotFoundException extends RuntimeException {
-
-    public NotFoundException(String message) {
-        super(message);
-    }
-}
