--- src/main/java/org/alpha/omega/hogwarts_artifacts_online/artifact/exception/error/FieldError.java
@@ -1,10 +1,0 @@
-package org.alpha.omega.hogwarts_artifacts_online.artifact.exception.error;
-
-import lombok.Builder;
-
-@Builder
-public record FieldError (
-        String field,
-        String message) {
-
-}
