+++ src/main/java/org/alpha/omega/hogwarts_artifacts_online/artifact/repository/ArtifactRepository.java
@@ -1,6 +1,6 @@
 package org.alpha.omega.hogwarts_artifacts_online.artifact.repository;
 
-import org.alpha.omega.hogwarts_artifacts_online.artifact.entity.Artifact;
+import org.alpha.omega.hogwarts_artifacts_online.entity.Artifact;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
