--- src/main/java/org/alpha/omega/hogwarts_artifacts_online/artifact/repository/WizardRepository.java
@@ -1,9 +1,0 @@
-package org.alpha.omega.hogwarts_artifacts_online.artifact.repository;
-
-import org.alpha.omega.hogwarts_artifacts_online.artifact.entity.Wizard;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface WizardRepository extends JpaRepository<Wizard, Long> {
-}
