--- src/main/java/org/alpha/omega/hogwarts_artifacts_online/artifact/response/ArtifactDTO.java
@@ -1,13 +1,0 @@
-package org.alpha.omega.hogwarts_artifacts_online.artifact.response;
-
-import lombok.Builder;
-
-@Builder
-public record ArtifactDTO(
-        String id,
-        String name,
-        String description,
-        String imageUrl,
-        WizardDTO owner
-) {
-}
