--- src/main/java/org/alpha/omega/hogwarts_artifacts_online/artifact/response/WizardDTO.java
@@ -1,11 +1,0 @@
-package org.alpha.omega.hogwarts_artifacts_online.artifact.response;
-
-import lombok.Builder;
-
-@Builder
-public record WizardDTO(
-        Long id,
-        String name,
-        Integer numberOfArtifacts
-) {
-}
