+++ src/main/java/org/alpha/omega/hogwarts_artifacts_online/artifact/service/ArtifactService.java
@@ -1,11 +1,11 @@
 package org.alpha.omega.hogwarts_artifacts_online.artifact.service;
 
 import lombok.RequiredArgsConstructor;
-import org.alpha.omega.hogwarts_artifacts_online.artifact.constant.Constant;
-import org.alpha.omega.hogwarts_artifacts_online.artifact.entity.Artifact;
-import org.alpha.omega.hogwarts_artifacts_online.artifact.exception.NotFoundException;
+import org.alpha.omega.hogwarts_artifacts_online.common.Constant;
+import org.alpha.omega.hogwarts_artifacts_online.entity.Artifact;
+import org.alpha.omega.hogwarts_artifacts_online.common.exception.NotFoundException;
 import org.alpha.omega.hogwarts_artifacts_online.artifact.repository.ArtifactRepository;
-import org.alpha.omega.hogwarts_artifacts_online.artifact.utility.IdWorker;
+import org.alpha.omega.hogwarts_artifacts_online.common.utility.IdWorker;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
