+++ src/main/java/org/alpha/omega/hogwarts_artifacts_online/common/exception/NotFoundException.java
@@ -1,0 +1,8 @@
+package org.alpha.omega.hogwarts_artifacts_online.common.exception;
+
+public class NotFoundException extends RuntimeException {
+
+    public NotFoundException(String message) {
+        super(message);
+    }
+}
