--- src/test/java/org/alpha/omega/hogwarts_artifacts_online/artifact/constant/TestConstant.java
@@ -1,21 +1,0 @@
-package org.alpha.omega.hogwarts_artifacts_online.artifact.constant;
-
-public class TestConstant {
-
-    private TestConstant() {}
-
-    public static final String ARTIFACT_ID = "10435344876";
-
-    public static class Exception {
-
-        private Exception() {}
-
-        public static class Artifact {
-
-            private Artifact() {}
-
-            public static final Object INVALID_ARGUMENTS = "Provided arguments are invalid, see data for details.";
-            public static final String NOT_FOUNT_ARTIFACT = "Could not find artifact with Id: %s";
-        }
-    }
-}
